Function Organization & Event Update
