+++ thirskOS_flutter/.packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2019-09-23 19:11:56.054540.
+# Generated by pub on 2019-11-05 12:38:48.297140.
 analyzer:file:///F:/flutter/.pub-cache/hosted/pub.dartlang.org/analyzer-0.34.3/lib/
 args:file:///F:/flutter/.pub-cache/hosted/pub.dartlang.org/args-1.5.1/lib/
 async:file:///F:/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.0.8/lib/
